pyproject.toml
[project]
name = "shortlab"
version = "0.0.0"
description = "Local deterministic pipeline for AI-generated Shorts"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "pycairo>=1.26",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "redis>=5.0",
  "rq>=1.16",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
